2020 Day 13 Part 2 'Solution: 1058443396696792' in 59.981µs

+++ go/2020/internal/gcd/gcd.go
@@ -1,7 +1,7 @@
 package gcd
 
 //GCD greatest common divisor (GCD) via Euclidean algorithm
-func GCD(a, b int) int {
+func GCD(a, b uint64) uint64 {
 	for b != 0 {
 		t := b
 		b = a % b
@@ -11,7 +11,7 @@ func GCD(a, b int) int {
 }
 
 //LCM find Least Common Multiple (LCM) via GCD
-func LCM(a, b int, integers ...int) int {
+func LCM(a, b uint64, integers ...uint64) uint64 {
 	result := a * b / GCD(a, b)
 
 	for i := 0; i < len(integers); i++ {
